+++ src/app/products/[auctionId]/_hooks/queries/useGetAuctionDetail.ts
