+++ src/app/products/new/_component/AuctionInfo/index.tsx
