+++ src/app/products/new/_component/InputLabel/index.tsx
