+++ src/app/products/new/_component/ProductInfo/ImageUploader/index.tsx
