+++ src/app/products/new/_component/ProductInfo/index.tsx
