+++ src/app/products/new/_hooks/mutations/useImageUpload.ts
