+++ src/app/products/new/_hooks/mutations/useReigsterProduct.ts
