+++ src/app/products/new/page.tsx
@@ -4,10 +4,10 @@ import { FormProvider, SubmitHandler, useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { z } from "zod";
 
-import AuctionInfo from "./_components/AuctionInfo";
-import ProductInfo from "./_components/ProductInfo";
-import { useImageUpload } from "@/app/hooks/mutations/useImageUpload";
-import { useRegisterProduct } from "@/app/hooks/mutations/useReigsterProduct";
+import AuctionInfo from "./_component/AuctionInfo";
+import ProductInfo from "./_component/ProductInfo";
+import { useImageUpload } from "@/app/products/new/_hooks/mutations/useImageUpload";
+import { useRegisterProduct } from "@/app/products/new/_hooks/mutations/useReigsterProduct";
 import { ProductSchema } from "./_utils/validation";
 import useBeforeUnload from "./_utils/useBeforeUnload";
 import { useRouter } from "next/navigation";
